vrisa_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.16,<3.17
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
PyJWT>=2.8.0,<3.0.0
